编辑任务开发

+++ src/main/java/org/nlpcn/jcoder/controller/TaskAction.java
@@ -5,10 +5,11 @@ import com.alibaba.fastjson.JSONObject;
 import com.google.common.collect.ImmutableMap;
 import org.h2.util.DateTimeUtils;
 import org.nlpcn.jcoder.constant.Api;
+import org.nlpcn.jcoder.constant.Constants;
 import org.nlpcn.jcoder.constant.TaskStatus;
 import org.nlpcn.jcoder.constant.TaskType;
+import org.nlpcn.jcoder.domain.HostGroup;
 import org.nlpcn.jcoder.domain.Task;
-import org.nlpcn.jcoder.domain.TaskHistory;
 import org.nlpcn.jcoder.domain.User;
 import org.nlpcn.jcoder.filter.AuthoritiesManager;
 import org.nlpcn.jcoder.run.CodeException;
@@ -27,9 +28,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Date;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 
 import static org.nlpcn.jcoder.constant.Constants.CURRENT_USER;
 import static org.nlpcn.jcoder.constant.Constants.TIMEOUT;
@@ -65,14 +64,20 @@ public class TaskAction {
      */
     @At
     public Restful list(String groupName, @Param(value = "taskType", df = "-1") int taskType) throws Exception {
+        List<HostGroup> hosts = new ArrayList<>(groupService.getGroupHostList(groupName));
         Object[] tasks = taskService.getTasksByGroupNameFromCluster(groupName)
                 .stream()
                 .filter(t -> taskType == -1 || Objects.equals(t.getType(), taskType))
-                .map(t -> ImmutableMap.of("name", t.getName(),
-                        "description", t.getDescription(),
-                        "status", t.getStatus(),
-                        "createTime", DateTimeUtils.formatDateTime(t.getCreateTime(), DATETIME_FORMAT, null, null),
-                        "updateTime", DateTimeUtils.formatDateTime(t.getUpdateTime(), DATETIME_FORMAT, null, null)))
+                .map(t -> {
+                    Map<String, Object> m = new HashMap<>(6);
+                    m.put("name", t.getName());
+                    m.put("description", t.getDescription());
+                    m.put("status", t.getStatus());
+                    m.put("hosts", hosts);
+                    m.put("createTime", DateTimeUtils.formatDateTime(t.getCreateTime(), DATETIME_FORMAT, null, null));
+                    m.put("updateTime", DateTimeUtils.formatDateTime(t.getUpdateTime(), DATETIME_FORMAT, null, null));
+                    return m;
+                })
                 .toArray();
         return Restful.instance(tasks);
     }
@@ -87,7 +92,7 @@ public class TaskAction {
      */
     @At
     public Restful save(@Param("hosts[]") String[] hosts, @Param("::task") Task task) throws Exception {
-        if (hosts == null) {
+        if (hosts == null || hosts.length < 1) {
             throw new IllegalArgumentException("empty hosts");
         }
 
@@ -123,7 +128,8 @@ public class TaskAction {
             return Restful.ERR.code(ServerException).msg(errorMessage);
         }
 
-        // ZK保存
+        // ZK保存, 给任务设置ID值, 方便后续做编辑
+        task.setId(0L);
         StaticValue.space().addTask(task);
 
         return Restful.OK;
@@ -148,59 +154,68 @@ public class TaskAction {
         return Restful.OK;
     }
 
-    @At("/task/editor/?/?")
-    @Ok("jsp:/task/task_editor.jsp")
-    public void find(Long groupId, Long taskId, @Param("version") String version) {
-        Mvcs.getReq().setAttribute("groupId", groupId);
-        if (!StringUtil.isBlank(version)) {
-            TaskHistory task = TaskService.findTaskByDBHistory(taskId, version);
-            Mvcs.getReq().setAttribute("task", task);
-            taskId = task.getTaskId();
-        } else {
-            Task task = TaskService.findTaskByDB(taskId);
-            Mvcs.getReq().setAttribute("task", new TaskHistory(task));
-            version = task.getVersion();
-            taskId = task.getId();
-        }
-
-        List<String> versions = taskService.versions(taskId, 100);
-        Mvcs.getReq().setAttribute("versions", versions);
-    }
-
-    @At("/task/find/?/?")
-    @Ok("json")
-    public JSONObject findApi(Long groupId, Long taskId, @Param("version") String version) {
-        JSONObject result = new JSONObject();
-        Mvcs.getReq().setAttribute("groupId", groupId);
-        if (!StringUtil.isBlank(version)) {
-            TaskHistory task = TaskService.findTaskByDBHistory(taskId, version);
-            result.put("task", task);
-            taskId = task.getTaskId();
-        } else {
-            Task task = TaskService.findTaskByDB(taskId);
-            result.put("task", new TaskHistory(task));
-            version = task.getVersion();
-            taskId = task.getId();
-        }
-        return result;
-    }
-
     /**
-     * 删除任务
-     * 如果任务类型是垃圾, 就物理删除
-     * 如果任务类型不是垃圾, 就更改任务类型和状态
+     * 获取任务
      *
-     * @param groupName
-     * @param name
+     * @param groupName  组名
+     * @param name       任务名
+     * @param sourceHost 来源主机, 即获取哪个主机的任务, 默认主版本
      * @return
      * @throws Exception
      */
     @At
-    public Restful delete(@Param("hosts[]") String[] hosts, String groupName, String name, int type) throws Exception {
-        if (hosts == null) {
-            throw new IllegalArgumentException("empty hosts");
+    public Restful task(String groupName, String name, @Param(value = "sourceHost", df = Constants.HOST_MASTER) String sourceHost) throws Exception {
+        if (StringUtil.isBlank(groupName)) {
+            throw new IllegalArgumentException("empty groupName");
         }
 
+        if (StringUtil.isBlank(name)) {
+            throw new IllegalArgumentException("empty name");
+        }
+
+        // 如果取主版本, 直接访问ZK
+        if (Constants.HOST_MASTER.equals(sourceHost)) {
+            Optional<Task> opt = taskService.getTasksByGroupNameFromCluster(groupName).stream().filter(t -> Objects.equals(t.getName(), name)).findAny();
+            if (!opt.isPresent()) {
+                LOG.warn("task[{}-{}] not found in zookeeper", groupName, name);
+                return Restful.OK;
+            }
+
+            return Restful.OK.obj(opt.get());
+        }
+
+        // 如果取其他机器版本
+        String content = proxyService.post(sourceHost, Api.TASK_TASK.getPath(), ImmutableMap.of("groupName", groupName, "name", name), TIMEOUT).getContent();
+        JSONObject res = JSONObject.parseObject(content);
+        return res.getBooleanValue("ok") ? Restful.OK.obj(res.get("obj")) : Restful.ERR.code(ServerException).msg(res.getString("message"));
+    }
+
+    @At
+    public Restful __task__(String groupName, String name) {
+        // 从本地数据库中找到任务
+        Task t = taskService.findTask(groupName, name);
+        if (t == null) {
+            LOG.warn("task[{}-{}] not found in host[{}]", groupName, name, StaticValue.getHostPort());
+            return Restful.OK;
+        }
+
+        return Restful.OK.obj(t);
+    }
+
+    /**
+     * 删除任务, 支持删除单台机器和所有机器(包括ZK)
+     * 如果任务类型是垃圾, 就物理删除
+     * 如果任务类型不是垃圾, 就更改任务类型和状态
+     *
+     * @param host      主机, 如果有, 就删除指定机器, 否则删除所有
+     * @param groupName
+     * @param name
+     * @param type
+     * @return
+     * @throws Exception
+     */
+    @At
+    public Restful delete(String host, String groupName, String name, int type) throws Exception {
         if (StringUtil.isBlank(groupName)) {
             throw new IllegalArgumentException("empty groupName");
         }
@@ -212,31 +227,39 @@ public class TaskAction {
         User u = (User) Mvcs.getHttpSession().getAttribute(CURRENT_USER);
         Date now = new Date();
 
+        // 主机列表
+        boolean isSingle = StringUtil.isNotBlank(host);
+        String[] hosts = isSingle ? new String[]{host} : groupService.getGroupHostList(groupName).stream().map(HostGroup::getHostPort).toArray(String[]::new);
+
         if (type == TaskType.RECYCLE.getValue()) {
             // 如果是垃圾, 物理删除
             // 删除集群的每台机器
-            String errorMessage = proxyService.post(hosts, Api.TASK_DELETE.getPath(), ImmutableMap.of("force", true, "groupName", groupName, "name", name, "user", u.getName(), "time", now), TIMEOUT, MERGE_FALSE_MESSAGE_CALLBACK);
+            String errorMessage = proxyService.post(hosts, Api.TASK_DELETE.getPath(), ImmutableMap.of("force", true, "groupName", groupName, "name", name, "user", u.getName(), "time", now.getTime()), TIMEOUT, MERGE_FALSE_MESSAGE_CALLBACK);
             if (StringUtil.isNotBlank(errorMessage)) {
                 return Restful.ERR.code(ServerException).msg(errorMessage);
             }
 
-            // 删ZK
-            taskService.deleteTaskFromCluster(groupName, name);
+            if (!isSingle) {
+                // 如果删除所有的, 删ZK
+                taskService.deleteTaskFromCluster(groupName, name);
+            }
         } else {
             // 更改类型为垃圾, 并停用
             // 更新集群的每台机器
-            String errorMessage = proxyService.post(hosts, Api.TASK_SAVE.getPath(), ImmutableMap.of("groupName", groupName, "name", name, "user", u.getName(), "time", now), TIMEOUT, MERGE_FALSE_MESSAGE_CALLBACK);
+            String errorMessage = proxyService.post(hosts, Api.TASK_DELETE.getPath(), ImmutableMap.of("groupName", groupName, "name", name, "user", u.getName(), "time", now.getTime()), TIMEOUT, MERGE_FALSE_MESSAGE_CALLBACK);
             if (StringUtil.isNotBlank(errorMessage)) {
                 return Restful.ERR.code(ServerException).msg(errorMessage);
             }
 
-            // 更新ZK
-            Task task = taskService.getTasksByGroupNameFromCluster(groupName).parallelStream().filter(t -> Objects.equals(t.getName(), name)).findAny().get();
-            task.setUpdateUser(u.getName());
-            task.setUpdateTime(now);
-            task.setType(TaskType.RECYCLE.getValue());
-            task.setStatus(TaskStatus.STOP.getValue());
-            StaticValue.space().addTask(task);
+            if (!isSingle) {
+                // 如果更新所有的, 更新ZK
+                Task task = taskService.getTasksByGroupNameFromCluster(groupName).parallelStream().filter(t -> Objects.equals(t.getName(), name)).findAny().get();
+                task.setUpdateUser(u.getName());
+                task.setUpdateTime(now);
+                task.setType(TaskType.RECYCLE.getValue());
+                task.setStatus(TaskStatus.STOP.getValue());
+                StaticValue.space().addTask(task);
+            }
         }
 
         return Restful.OK;
@@ -260,6 +283,11 @@ public class TaskAction {
 
         // 从本地数据库中找到任务
         Task t = taskService.findTask(groupName, name);
+        if (t == null) {
+            LOG.warn("task[{}-{}] not found in host[{}]", groupName, name, StaticValue.getHostPort());
+            return Restful.OK;
+        }
+
         t.setUpdateUser(user);
         t.setUpdateTime(time);
         taskService.delete(t);
